refactor(Dock): 重构Dock功能模块结构，将相关代码迁移至Features目录
将Dock相关代码从Settings和ShortcutDock目录迁移至Features/Dock目录，包括：
- Dock设置模型和默认值
- 快捷入口启动帮助类
- 图标加载服务
- 主窗口Dock区域UI引用
- Dock功能编排逻辑
- 相关单元测试

同时更新了各文件的命名空间引用以匹配新位置
