+++ WindBoard/Features/Dock/UI/DockSettingsPage.xaml
@@ -1,11 +1,11 @@
 <?xml version="1.0" encoding="utf-8"?>
 <Page
-    x:Class="WindBoard.Settings.Pages.DockSettingsPage"
+    x:Class="WindBoard.Features.Dock.UI.DockSettingsPage"
     xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
     xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
     xmlns:l10n="using:WindBoard.Localization"
     xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
-    xmlns:local="using:WindBoard.Settings.Pages"
+    xmlns:local="using:WindBoard.Features.Dock.UI"
     xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
     mc:Ignorable="d">
