+++ WindBoard/Settings/AppSettings.cs
@@ -1,5 +1,6 @@
 using System;
 using System.Collections.Generic;
+using WindBoard.Features.Dock.Models;
 
 namespace WindBoard.Settings
 {
