+++ WindBoard/Settings/AppSettingsCloner.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using WindBoard.Features.Dock.Models;
 
 namespace WindBoard.Settings
 {
