+++ WindBoard/Settings/AppSettingsService.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.Threading;
 using System.Threading.Tasks;
 using Windows.UI;
+using WindBoard.Features.Dock.Models;
 using WindBoard.Logging;
 
 namespace WindBoard.Settings
