+++ WindBoard/Settings/AppSettingsStore.cs
@@ -2,6 +2,7 @@ using System;
 using System.Collections.Generic;
 using System.IO;
 using System.Text.Json;
+using WindBoard.Features.Dock.Models;
 using WindBoard.Logging;
 using WindBoard.Persistence;
